+++ test/run_all_versions.sh
@@ -104,6 +104,7 @@ SUITES=(
 	import_export_privilege
 	index_only
 	isolation
+	local_open_race_free
 	logical_subscriber
 	maintenance_due
 	native_agg